added unit test for utils.readDateTime
